+++ src/design-system/core/index.ts
@@ -1,3 +1,4 @@
 export * from './components'
+export * from './i18n'
 export * from './icons'
 export * from './schema'