+++ src/design-system/vue/index.ts
@@ -1,3 +1,4 @@
 export * from './components'
 export * from './compatibility'
+export * from './composables/useNtkI18n'
 export * from './plugin'